made wrapper component for editor
