+++ frontends/main/src/app-pages/Articles/ArticleEditPage.test.tsx
@@ -13,33 +13,6 @@ jest.mock("next-nprogress-bar", () => ({
   }),
 }))
 
-jest.mock("ol-components", () => {
-  // Reuse other exports from ol-components if needed
-  const actual = jest.requireActual("ol-components")
-  return {
-    ...actual,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    TiptapEditor: ({ value, onChange, "data-testid": testId }: any) => {
-      return (
-        <textarea
-          data-testid={testId || "editor"}
-          value={JSON.stringify(value, null, 2)}
-          onChange={(e) => {
-            try {
-              // Allow simulating JSON-like updates if needed
-              const parsed = JSON.parse(e.target.value)
-              onChange?.(parsed)
-            } catch {
-              // fallback to raw string for simple tests
-              onChange?.(e.target.value)
-            }
-          }}
-        />
-      )
-    },
-  }
-})
-
 describe("ArticleEditPage", () => {
   test("renders editor when user has ArticleEditor permission", async () => {
     const user = factories.user.user({
@@ -51,7 +24,15 @@ describe("ArticleEditPage", () => {
     const article = factories.articles.article({
       id: 42,
       title: "Existing Title",
-      content: { id: 1, content: "Existing content" },
+      content: {
+        type: "doc",
+        content: [
+          {
+            type: "paragraph",
+            content: [{ type: "text", text: "Existing Title" }],
+          },
+        ],
+      },
     })
     setMockResponse.get(urls.articles.details(article.id), article)
 
@@ -72,7 +53,15 @@ describe("ArticleEditPage", () => {
     const article = factories.articles.article({
       id: 123,
       title: "Existing Title",
-      content: { id: 1, content: "Existing content" },
+      content: {
+        type: "doc",
+        content: [
+          {
+            type: "paragraph",
+            content: [{ type: "text", text: "Existing Title" }],
+          },
+        ],
+      },
     })
     setMockResponse.get(urls.articles.details(article.id), article)
 
@@ -104,7 +93,15 @@ describe("ArticleEditPage", () => {
     const article = factories.articles.article({
       id: 7,
       title: "Old Title",
-      content: { id: 1, content: "Bad content" },
+      content: {
+        type: "doc",
+        content: [
+          {
+            type: "paragraph",
+            content: [{ type: "text", text: "Existing Title" }],
+          },
+        ],
+      },
     })
     setMockResponse.get(urls.articles.details(article.id), article)
