+++ frontends/main/src/app-pages/Articles/ArticleEditPage.tsx
@@ -3,15 +3,17 @@ import React, { useEffect, useState } from "react"
 import { Permission } from "api/hooks/user"
 import { useRouter } from "next-nprogress-bar"
 import { useArticleDetail, useArticlePartialUpdate } from "api/hooks/articles"
-import { Button, Input, Alert } from "@mitodl/smoot-design"
+import { Button, Alert } from "@mitodl/smoot-design"
 import RestrictedRoute from "@/components/RestrictedRoute/RestrictedRoute"
 import {
   Container,
   Typography,
   styled,
   LoadingSpinner,
-  TiptapEditor,
+  TiptapEditorContainer,
+  JSONContent,
 } from "ol-components"
+
 import { notFound } from "next/navigation"
 import { articlesView } from "@/common/urls"
 
@@ -25,18 +27,6 @@ const ClientContainer = styled.div({
   margin: "10px 0",
 })
 
-const TitleInput = styled(Input)({
-  width: "100%",
-  margin: "10px 0",
-})
-
-styled(TiptapEditor)({
-  ".simple-editor-wrapper": {
-    width: "auto",
-    height: "auto",
-  },
-})
-
 const ArticleEditPage = ({ articleId }: { articleId: string }) => {
   const router = useRouter()
 
@@ -44,8 +34,7 @@ const ArticleEditPage = ({ articleId }: { articleId: string }) => {
   const { data: article, isLoading } = useArticleDetail(id)
 
   const [title, setTitle] = useState<string>("")
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const [json, setJson] = useState<Record<string, any>>({
+  const [json, setJson] = useState<JSONContent>({
     type: "doc",
     content: [{ type: "paragraph", content: [] }],
   })
@@ -91,7 +80,7 @@ const ArticleEditPage = ({ articleId }: { articleId: string }) => {
 
   return (
     <RestrictedRoute requires={Permission.ArticleEditor}>
-      <Container className="article-wrapper">
+      <Container>
         <Typography variant="h3" component="h1">
           Edit Article
         </Typography>
@@ -107,23 +96,17 @@ const ArticleEditPage = ({ articleId }: { articleId: string }) => {
             </Typography>
           </Alert>
         )}
-        <TitleInput
-          type="text"
-          value={title}
-          onChange={(e) => {
-            console.log("Title input changed:", e.target.value)
-            setTitle(e.target.value)
-            setAlertText("")
-          }}
-          placeholder="Enter article title"
-          className="input-field"
-        />
 
-        <ClientContainer className="editor-box">
-          <TiptapEditor
+        <ClientContainer>
+          <TiptapEditorContainer
             data-testid="editor"
             value={json}
             onChange={handleChange}
+            title={title}
+            setTitle={(e) => {
+              setTitle(e.target.value)
+              setAlertText("")
+            }}
           />
         </ClientContainer>
