+++ frontends/main/src/app-pages/Articles/NewArticlePage.test.tsx
@@ -17,33 +17,6 @@ jest.mock("next/navigation", () => ({
   }),
 }))
 
-jest.mock("ol-components", () => {
-  // Reuse other exports from ol-components if needed
-  const actual = jest.requireActual("ol-components")
-  return {
-    ...actual,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    TiptapEditor: ({ value, onChange, "data-testid": testId }: any) => {
-      return (
-        <textarea
-          data-testid={testId || "editor"}
-          value={JSON.stringify(value, null, 2)}
-          onChange={(e) => {
-            try {
-              // Allow simulating JSON-like updates if needed
-              const parsed = JSON.parse(e.target.value)
-              onChange?.(parsed)
-            } catch {
-              // fallback to raw string for simple tests
-              onChange?.(e.target.value)
-            }
-          }}
-        />
-      )
-    },
-  }
-})
-
 describe("ArticleNewPage", () => {
   test("throws ForbiddenError when user lacks ArticleEditor permission", async () => {
     const user = factories.user.user({
